requirements.txt
fastapi
uvicorn[standard]
pydantic
lightgbm
pandas
geopandas
psycopg2-binary
shap
joblib
python-jose[cryptography]   # JWT
